fix: fixed bugs

+++ libs.h
@@ -7,6 +7,8 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <string.h>
+#include <stdbool.h>
+#include <fcntl.h>
 
 #include <linux/blkpg.h>
 #include <linux/magic.h>
@@ -17,6 +19,6 @@
 #include <sys/statvfs.h>
 
 /*A function to proccessing/get a file to/from an empty place in the cluster*/
-int8_t save_file(int64_t, int64_t);
+int8_t save_file(int32_t, int32_t);
 
-int8_t get_file(int64_t, int64_t);
+int8_t get_file(int32_t, int32_t);
