+++ src/get_file.c
@@ -1,7 +1,7 @@
 // Copyright 2024 binoll
 #include "../libs.h"
 
-int8_t get_file(int64_t fd, int64_t cd) {
+int8_t get_file(int32_t fd, int32_t cd) {
 
 	return EXIT_SUCCESS;
 }
